Drop useless u64 conversions in the Darwin start-token reader.
proc_bsdinfo start-time fields are already u64 on macOS, so the
conversions are identity and clippy rejects them under -D warnings in
the macOS validate job. The checked arithmetic is unchanged.

+++ crates/mez-mux/src/process/process_metadata.rs
@@ -186,9 +186,9 @@ pub fn process_start_token_for_pid(pid: u32) -> Option<u64> {
     }
     // SAFETY: the exact structure size was initialized successfully above.
     let info = unsafe { info.assume_init() };
-    u64::from(info.pbi_start_tvsec)
+    info.pbi_start_tvsec
         .checked_mul(1_000_000)?
-        .checked_add(u64::from(info.pbi_start_tvusec))
+        .checked_add(info.pbi_start_tvusec)
 }
 
 /// Returns no creation-time token on hosts without a reviewed native reader.
